Compile on ARM platform
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@18673 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ src/KDChartMeasure.cpp
@@ -187,7 +187,7 @@ bool Measure::operator==( const Measure& r ) const
 
 GlobalMeasureScaling::GlobalMeasureScaling()
 {
-    mFactors.push( qMakePair(1.0, 1.0) );
+    mFactors.push( qMakePair(qreal(1.0), qreal(1.0)) );
 }
 
 GlobalMeasureScaling::~GlobalMeasureScaling()
